Fixing the about and categories pages
