+++ Frontend/src/App.tsx
@@ -1,20 +1,18 @@
 import { Route, Routes } from "react-router-dom";
 
 import IndexPage from "@/pages/index";
-import DocsPage from "@/pages/docs";
-import PricingPage from "@/pages/pricing";
 import BlogPage from "@/pages/blog";
-import AboutPage from "@/pages/about";
 import LoginPage from "./pages/login";
+import AboutPage from "./pages/about";
+import CategoriesPage from "@/pages/categories";
 
 function App() {
   return (
     <Routes>
       <Route element={<IndexPage />} path="/" />
-      <Route element={<DocsPage />} path="/docs" />
-      <Route element={<PricingPage />} path="/pricing" />
-      <Route element={<BlogPage />} path="/blog" />
+      <Route element={<CategoriesPage />} path="/categories" />
       <Route element={<AboutPage />} path="/about" />
+      <Route element={<BlogPage />} path="/blog" />
       <Route element={<LoginPage/>}  path="/login"/>
     </Routes>
   );
